Reorganize ProfileViewModel and ProfileEditingViewModel to force consistency when  checking for changes
