+++ Mastodon/In Progress New Layout and Datamodel/Profile/ProfileEditingView.swift
@@ -274,10 +274,7 @@ struct ProfileEditingView: View {
     }
 }
 
-struct ORIGINALProfileEditingView: View {
-    @Environment(ProfileViewModel.self) var profileViewModel
-    @Environment(ProfileEditingViewModel.self) var editingViewModel
-    
+struct ORIGINALProfileEditingView {
     enum FieldEditType {
         case create
         case edit(Mastodon.Entity.Field)
@@ -296,488 +293,6 @@ struct ORIGINALProfileEditingView: View {
         let editingField: FieldEditType
         let labelEditingModel: MetaTextInputFieldViewModel
         let valueEditingModel: MetaTextInputFieldViewModel
-    }
-
-    var body: some View {
-        GeometryReader { geo in
-            ScrollView {
-                VStack(alignment: .leading, spacing: 0) {
-                    ProfileAvatarAndBannerView(width: geo.size.width)
-                    
-                    nameAndBio
-                        .padding(.horizontal)
-                        .padding(.vertical, doublePadding)
-                    
-                    Divider()
-                    
-                    CustomProfileFieldsEditor()
-                        .padding(.horizontal)
-                        .padding(.vertical, doublePadding)
-                    
-                    Divider()
-                    
-                    if editingViewModel.showTabDisplayPreferences {
-                        ProfileTabSettingsEditor()
-                            .padding(.horizontal)
-                            .padding(.vertical, doublePadding)
-                        
-                        Divider()
-                    }
-                    
-                    // ADVANCED SETTINGS
-                    
-                    AdvancedSettingsEditor()
-                        .padding(.horizontal)
-                        .padding(.vertical, doublePadding)
-                    
-                    Spacer()
-                        .frame(maxHeight: .infinity)
-                }
-                .overlay {
-                    if editingViewModel.bioFieldEditingViewModel.isEditing {
-                        editingViewModel.bioFieldEditingViewModel.autoCompleteSuggestionView(pinToTopOfKeyboard: true)
-                    } else if profileViewModel.editingStatus.showActivityIndicator {
-                        ZStack {
-                            Color.dimmingBackground
-                            ProgressView()
-                                .progressViewStyle(.circular)
-                        }
-                    }
-                }
-                .sheet(isPresented: editingViewModel.showCroppingView) {
-                    if let image = editingViewModel.avatarImageToCrop {
-                        PhotoCropperView(originalImage: image) { confirmedImage in
-                            defer {
-                                self.editingViewModel.avatarImageToCrop = nil
-                            }
-                            
-                            guard let confirmedImage else {
-                                self.editingViewModel.avatarConfirmedCroppedImage = nil
-                                return
-                            }
-                            
-                            self.editingViewModel.avatarConfirmedCroppedImage = confirmedImage
-                        }
-                    }
-                }
-            }
-        }
-        .navigationTitle(L10nLookup.Scene.EditProfile.title)
-        .toolbar {
-            saveButtonForToolbar
-        }
-        .overlay() {
-            if let fieldEditingState = editingViewModel.fieldEditingState {
-                editFieldView(fieldEditingState)
-            }
-        }
-        .alert(editingViewModel.presentingAlert?.title ?? "",
-               isPresented: Binding<Bool>(
-                get: { editingViewModel.presentingAlert != nil },
-                set: { newValue in
-                    if newValue == false {
-                        editingViewModel.presentingAlert = nil
-                    }
-                }
-               ),
-               actions: {
-            Button(role: .cancel) {
-                withAnimation {
-                    editingViewModel.cancelDeleteCustomField()
-                }
-            } label: {
-                Text("Cancel")
-            }
-            Button(role: .destructive) {
-                withAnimation {
-                    editingViewModel.confirmDeleteCustomField()
-                }
-            } label: {
-                Text("Delete")
-            }
-        },
-               message: {
-            if let messageText = editingViewModel.presentingAlert?.messageText {
-                Text(messageText)
-            }
-        })
-        .onChange(of: editingViewModel.isAutomatedAccount) {
-            editingViewModel.checkForChanges()
-        }
-        .onChange(of: editingViewModel.avatarConfirmedCroppedImage) {
-            editingViewModel.checkForChanges()
-        }
-        .onChange(of: editingViewModel.confirmedBannerImage) {
-            editingViewModel.checkForChanges()
-        }
-        .onChange(of: editingViewModel.mediaTabVisibilitySetting) {
-            editingViewModel.checkForChanges()
-        }
-        .onChange(of: editingViewModel.mediaTabRepliesSetting) {
-            editingViewModel.checkForChanges()
-        }
-        .onChange(of: editingViewModel.featuredTabVisibilitySetting) {
-            editingViewModel.checkForChanges()
-        }
-    }
-    
-    @ViewBuilder var nameAndBio: some View {
-        VStack(spacing: doublePadding) {
-            VStack(alignment: .leading, spacing: tinySpacing) {
-                SubsectionHeading(title: "Display name", subtitle: nil) // TODO: needs L10n
-                MetaTextInputField(allowScroll: false, drawBackground: true, returnKeyType: .done)
-                    .environment(editingViewModel.displayNameFieldEditingViewModel)
-                    .frame(height: 36)
-            }
-            
-            VStack(alignment: .leading, spacing: tinySpacing) {
-                SubsectionHeading(title: "Bio", subtitle: "Introduce yourself. Recommended 220 character maximum.") // TODO: needs L10n
-                MetaTextInputField(allowScroll: true, drawBackground: true, returnKeyType: .done)
-                    .environment(editingViewModel.bioFieldEditingViewModel)
-                    .frame(height: 56)
-            }
-        }
-    }
-    
-    @ViewBuilder func editFieldView(_ editState: ORIGINALProfileEditingView.FieldEditingState) -> some View {
-        GeometryReader { _ in
-            ZStack {
-                Color.dimmingBackground
-                    .onTapGesture {
-                        self.editingViewModel.fieldEditingState = nil
-                    }
-                
-                VStack(alignment: .leading, spacing: doublePadding) {
-                    Text(editState.editingField.title)
-                        .fontWeight(.semibold)
-                    
-                    Divider()
-                    
-                    VStack(alignment: .leading, spacing: tinySpacing) {
-                        SubsectionHeading(title: "Label", subtitle: nil)
-                        MetaTextInputField(allowScroll: false, drawBackground: true, returnKeyType: .done)
-                            .environment(editState.labelEditingModel)
-                            .frame(height: 36)
-                    }
-                    
-                    VStack(alignment: .leading, spacing: tinySpacing) {
-                        SubsectionHeading(title: "Value", subtitle: nil)
-                        MetaTextInputField(allowScroll: true, drawBackground: true, returnKeyType: .done)
-                            .environment(editState.valueEditingModel)
-                            .frame(height: 36)
-                    }
-                    
-                    HStack {
-                        Spacer()
-                            .frame(maxWidth: .infinity)
-                        
-                        Button() {
-                            withAnimation {
-                                editingViewModel.fieldEditingState = nil
-                            }
-                        } label: {
-                            Text("Cancel")
-                                .padding(.horizontal)
-                                .padding(.vertical, tinySpacing)
-                                .background() {
-                                    Capsule()
-                                        .stroke(.secondary)
-                                }
-                        }
-                        
-                        Button() {
-                            withAnimation {
-                                editingViewModel.commitEditingField()
-                            }
-                        } label: {
-                            Text("Save") // TODO: L10n
-                                .foregroundColor(.white)
-                                .padding(.horizontal)
-                                .padding(.vertical, tinySpacing)
-                                .background() {
-                                    Capsule()
-                                        .fill(Asset.Colors.accent.swiftUIColor)
-                                }
-                        }
-                    }
-                    .fontWeight(.semibold)
-                }
-                .frame(maxWidth: 300)
-                .padding(doublePadding)
-                .background() {
-                    RoundedRectangle(cornerRadius: CornerRadius.extraLarge)
-                        .fill(.background)
-                        .frame(maxWidth: .infinity, maxHeight: .infinity)
-                }
-            }
-        }
-    }
-    
-    @ViewBuilder func alertContents(_ alert: ProfileEditingViewModel.ProfileEditingAlert) -> some View {
-        Color.red
-            .frame(width: 200, height: 50)
-    }
-    
-    @ToolbarContentBuilder var saveButtonForToolbar: some ToolbarContent {
-        switch profileViewModel.editingStatus.saveButton {
-        case .canSave:
-            ToolbarItem(placement: .navigationBarTrailing) {
-                Button(L10n.Common.Controls.Actions.save) {
-                    Task {
-                        try await profileViewModel.commitEdits()
-                    }
-                }
-                .buttonStyle(.borderedProminent)
-                .tint(Asset.Colors.accent.swiftUIColor)
-            }
-        default:
-            ToolbarItem(placement: .navigationBarTrailing){}
-        }
-    }
-}
-
-@MainActor
-@Observable
-class ProfileEditingViewModel {
-    var editingStatus: EditingStatus?
-    var showVerifiedLinkTip = true
-    var showTabDisplayPreferences = false
-    
-    let displayNameFieldEditingViewModel = MetaTextInputFieldViewModel(stringContent: "", placeholder: "", characterLimit: .hardLimit(30), autocompleteMastodonItems: false)
-    let bioFieldEditingViewModel = MetaTextInputFieldViewModel(stringContent: "", placeholder: "Describe yourself and/or this account.", characterLimit: .softLimit(220), autocompleteMastodonItems: true) // TODO: L10n
-    
-    var fieldEditingState: ORIGINALProfileEditingView.FieldEditingState?
-    var isReorderingCustomFields: Bool = false
-    
-    var selectedBannerImage: Binding<[PhotosPickerItem]>
-    var bannerImagePhotosPickerItem: PhotosPickerItem?
-    var confirmedBannerImage: UIImage?
-    
-    var selectedAvatar: Binding<[PhotosPickerItem]>
-    var avatarPhotosPickerItem: PhotosPickerItem?
-    var avatarImageToCrop: UIImage?
-    var avatarConfirmedCroppedImage: UIImage?
-    
-    var showCroppingView: Binding<Bool>
-    var presentingAlert: ProfileEditingAlert?
-    
-    private(set) var mediaTabVisibilitySetting: MediaTabVisibilitySetting = .showMediaTab
-    private(set) var mediaTabRepliesSetting: MediaTabRepliesSetting = .showDirectPostsOnly
-    private(set) var featuredTabVisibilitySetting: FeaturedTabVisibilitySetting = .showFeaturedTab
-    
-    var customFields: [Mastodon.Entity.Field]? = nil
-    var reorderingCustomFields: [Mastodon.Entity.Field] = []
-    var emojis: [Mastodon.Entity.Emoji] = []
-    
-    var originalFeaturedTags: [Mastodon.Entity.FeaturedTag] = []
-    var editedFeaturedTags: [Mastodon.Entity.FeaturedTag] = []
-    
-    var isAutomatedAccount: Bool = false
-    
-    private(set) var initialInfo: MastodonAccount? = nil
-    
-    init() {
-        selectedBannerImage = Binding<[PhotosPickerItem]>(get: {[]}, set: {_ in})
-        selectedAvatar = Binding<[PhotosPickerItem]>(get: {[]}, set: {_ in})
-        showCroppingView = Binding<Bool>(get: {false}, set: {_ in})
-        
-        showCroppingView = Binding<Bool>(
-            get: { return self.avatarImageToCrop != nil },
-            set: { newValue in }
-        )
-        
-        selectedBannerImage = Binding<[PhotosPickerItem]>(
-            get: { [self.bannerImagePhotosPickerItem].compactMap { $0 } },
-            set: { newValue in
-                self.bannerImagePhotosPickerItem = newValue.first
-                if let item = self.bannerImagePhotosPickerItem {
-                    Task {
-                        if let data = try? await item.loadTransferable(type: Data.self),
-                           let image = UIImage(data: data) {
-                            self.confirmedBannerImage = image
-                        } else {
-                            print("Failed to load banner image")
-                        }
-                    }
-                }
-            }
-        )
-        
-        selectedAvatar = Binding<[PhotosPickerItem]>(
-            get: { [self.avatarPhotosPickerItem].compactMap { $0 } },
-            set: { newValue in
-                self.avatarPhotosPickerItem = newValue.first
-                if let item = self.avatarPhotosPickerItem {
-                    Task {
-                        if let data = try? await item.loadTransferable(type: Data.self),
-                           let image = UIImage(data: data) {
-                            self.avatarImageToCrop = image
-                        } else {
-                            print("Failed to load avatar image")
-                        }
-                    }
-                }
-            }
-        )
-        
-        displayNameFieldEditingViewModel.contentDidChange = { withAnimation { self.checkForChanges() } }
-        bioFieldEditingViewModel.contentDidChange = { withAnimation { self.checkForChanges() } }
-    }
-    
-    func checkForChanges() {
-        let hasChanges = confirmedBannerImage != nil ||
-        avatarConfirmedCroppedImage != nil ||
-        (initialInfo != nil && isAutomatedAccount != initialInfo?.metadata.isBot) ||
-        customFields != initialInfo?.metadata.customFieldsForEdit || displayNameFieldEditingViewModel.stringContent != displayNameFieldEditingViewModel.originalStringContent || bioFieldEditingViewModel.stringContent != bioFieldEditingViewModel.originalStringContent
-        if hasChanges {
-            withAnimation { editingStatus = .editing(hasChanges: true) }
-        }
-    }
-    
-    func setAccount(_ account: MastodonAccount) {
-        initialInfo = account
-        updateAccountTextFields(account: account)
-        updateCustomFields(account: account)
-        updateMetaData(account: account)
-        if customFields?.first(where: { $0.verifiedAt != nil }) != nil {
-            showVerifiedLinkTip = false
-        }
-        avatarConfirmedCroppedImage = nil
-        avatarPhotosPickerItem = nil
-        avatarImageToCrop = nil
-        confirmedBannerImage = nil
-        bannerImagePhotosPickerItem = nil
-        presentingAlert = nil
-        
-        showTabDisplayPreferences = AuthenticationServiceProvider.shared.currentActiveUser.value?.authentication.instanceConfiguration?.isAvailable(.profileSettings) ?? false
-    }
-    
-    func updateFeaturedHashtags(_ featuredTags: [Mastodon.Entity.FeaturedTag]) {
-        originalFeaturedTags = featuredTags
-        editedFeaturedTags = featuredTags
-    }
-    
-    func updateAccountTextFields(account: MastodonAccount) {
-        let displayNameContent = account.displayInfo.displayName
-        displayNameFieldEditingViewModel.stringContent = displayNameContent
-        
-        let bioContent = normalize(htmlString: account.bioForEdit)
-        bioFieldEditingViewModel.stringContent = bioContent ?? ""
-    }
-    
-    func updateCustomFields(account: MastodonAccount) {
-        customFields = account.metadata.customFieldsForEdit
-        emojis = account._legacyEntity.emojis
-    }
-    
-    func updateMetaData(account: MastodonAccount) {
-        mediaTabVisibilitySetting = account.metadata.showsMediaTab ? .showMediaTab : .hideMediaTab
-        mediaTabRepliesSetting = account.metadata.mediaTabIncludesReplies ? .includeMyRepliesToOthers : .showDirectPostsOnly
-        featuredTabVisibilitySetting = account.metadata.showsFeaturedTab ? .showFeaturedTab : .hideFeaturedTab
-        isAutomatedAccount = account.metadata.isBot
-    }
-    
-    func beginEditingField(_ fieldType: ORIGINALProfileEditingView.FieldEditType, profileViewModel: ProfileViewModel, navigator: MastodonNavigationRouter) {
-        switch fieldType {
-            // TODO: L10n for all placeholders
-        case .create:
-            fieldEditingState = .init(editingField: fieldType,
-                                      labelEditingModel: MetaTextInputFieldViewModel(stringContent: "", placeholder: "E.g. \"Personal website\"", characterLimit: .softLimit(100), autocompleteMastodonItems: false),
-                                      valueEditingModel: MetaTextInputFieldViewModel(stringContent: "", placeholder: "E.g. \"example.me\"", characterLimit: .softLimit(220), autocompleteMastodonItems: true))
-        case .edit(let field):
-            fieldEditingState = .init(editingField: fieldType,
-                                      labelEditingModel: MetaTextInputFieldViewModel(stringContent: field.name, placeholder: "E.g. \"Personal website\"", characterLimit: .softLimit(100), autocompleteMastodonItems: false),
-                                      valueEditingModel: MetaTextInputFieldViewModel(stringContent: field.value, placeholder: "E.g. \"example.me\"", characterLimit: .softLimit(220), autocompleteMastodonItems: true))
-        }
-        navigator.presentModal(.editProfileNavigation(destination: .editCustomField(profileViewModel: profileViewModel)))
-    }
-    
-    func beginReorderingFields(profileViewModel: ProfileViewModel, navigator: MastodonNavigationRouter) {
-        isReorderingCustomFields = true
-        reorderingCustomFields = customFields ?? []
-        navigator.presentModal(.editProfileNavigation(destination: .reorderCustomFields(profileViewModel: profileViewModel)))
-    }
-    
-    func deleteCurrentEditingField() {
-        guard let fieldEditingState else { return }
-        switch fieldEditingState.editingField {
-        case .create:
-            return
-        case .edit(let field):
-            self.fieldEditingState = nil
-            guard let index = customFields?.firstIndex(of: field) else { return }
-            customFields?.remove(at: index)
-        }
-    }
-    
-    func commitEditingField() {
-        guard let fieldEditingState else { return }
-        let labelText = fieldEditingState.labelEditingModel.stringContent
-        let valueText = fieldEditingState.valueEditingModel.stringContent
-        guard !labelText.isEmpty, !valueText.isEmpty else { return }
-        
-        let newField = Mastodon.Entity.Field(name: labelText, value: valueText)
-        
-        switch fieldEditingState.editingField {
-        case .create:
-            customFields?.append(newField)
-        case .edit(let field):
-            if let replaceIndex = customFields?.firstIndex(of: field) {
-                customFields?[replaceIndex] = newField
-            } else {
-                customFields?.append(newField)
-            }
-        }
-        self.fieldEditingState = nil
-        
-        checkForChanges()
-    }
-    
-    func commitCustomFieldsReordering() {
-        guard isReorderingCustomFields else { return }
-        customFields = reorderingCustomFields
-        isReorderingCustomFields = false
-        checkForChanges()
-    }
-    
-    func requestDeleteCustomField(_ field: Mastodon.Entity.Field) {
-        presentingAlert = .deleteCustomField(field)
-    }
-    
-    func confirmDeleteCustomField() {
-        let deletingField: Mastodon.Entity.Field? = {
-            switch presentingAlert {
-            case .deleteCustomField(let field):
-                return field
-            case nil:
-                return nil
-            }
-        }()
-        guard let field = deletingField, let index = customFields?.firstIndex(of: field) else { return }
-        customFields?.remove(at: index)
-        presentingAlert = nil
-        checkForChanges()
-    }
-    
-    func cancelDeleteCustomField() {
-        presentingAlert = nil
-    }
-}
-
-extension ProfileEditingViewModel {
-    func setMediaTabVisibilitySetting(_ newSetting: MediaTabVisibilitySetting) {
-        guard newSetting != mediaTabVisibilitySetting else { return }
-        mediaTabVisibilitySetting = newSetting
-    }
-    
-    func setMediaTabRepliesSetting(_ newSetting: MediaTabRepliesSetting) {
-        guard newSetting != mediaTabRepliesSetting else { return }
-        mediaTabRepliesSetting = newSetting
-    }
-    
-    func setFeaturedTabVisibilitySetting(_ newSetting: FeaturedTabVisibilitySetting) {
-        guard newSetting != featuredTabVisibilitySetting else { return }
-        featuredTabVisibilitySetting = newSetting
     }
 }
 
@@ -959,78 +474,6 @@ struct RadioButtonArray: View {
     }
 }
 
-let brandBackgroundColor: Color = Asset.Colors.Brand.lightBlurple.swiftUIColor.opacity(0.2)
-
-extension ProfileViewModel {
-    func commitEdits() async throws {
-        guard let authentication = AuthenticationServiceProvider.shared.currentActiveUser.value?.authentication else { throw APIService.APIError.explicit(.authenticationMissing) }
-        
-        let authenticationBox = MastodonAuthenticationBox(authentication: authentication)
-        let domain = authenticationBox.domain
-        let authorization = authenticationBox.userAuthorization
-        
-        func sizeLimitedImage(_ image: UIImage?, noLargerThan targetPixelSize: CGSize) -> UIImage? {
-            guard let image else { return nil }
-            if image.size.width <= targetPixelSize.width && image.size.height <= targetPixelSize.height {
-                return image
-            } else {
-                let resized = image.af.imageScaled(to: targetPixelSize)
-                return resized
-            }
-        }
-        
-        let updatedAvatarImage = sizeLimitedImage(editingViewModel.avatarConfirmedCroppedImage, noLargerThan: avatarImageMaxSizeInPixels)
-        let updatedBannerImage = sizeLimitedImage(editingViewModel.confirmedBannerImage, noLargerThan: bannerImageMaxSizeInPixels)
-        
-        editingViewModel.commitEditingField()
-        editingViewModel.commitCustomFieldsReordering()
-        let customFieldsData = editingViewModel.customFields?.map { Mastodon.Entity.Field(name: $0.name, value: $0.value) }
-        
-        let query = Mastodon.API.Account.UpdateCredentialQuery(
-            discoverable: nil,
-            bot: editingViewModel.isAutomatedAccount,
-            displayName: String(editingViewModel.displayNameFieldEditingViewModel.stringContent.prefix(30)),
-            note: editingViewModel.bioFieldEditingViewModel.stringContent,
-            avatar: updatedAvatarImage.flatMap { Mastodon.Query.MediaAttachment.png($0.pngData()) },
-            header: updatedBannerImage.flatMap { Mastodon.Query.MediaAttachment.png($0.pngData()) },
-            locked: nil,
-            source: nil,
-            fieldsAttributes: customFieldsData
-        )
-        
-        editingStatus = .pushingChanges(success: nil)
-        
-        do {
-            let response = try await APIService.shared.accountUpdateCredentials(
-                domain: domain,
-                query: query,
-                authorization: authorization
-            )
-            let updatedAccount = MastodonAccount.fromEntity(response.value, authenticatedDomain: domain)
-            account = updatedAccount
-            editingViewModel.setAccount(updatedAccount)
-            editingStatus = .pushingChanges(success: true)
-        } catch {
-            editingStatus = .pushingChanges(success: false)
-            throw error
-        }
-    }
-    
-    func commitTabSettingsChanges() async throws {
-        guard let navigator, let authBox = AuthenticationServiceProvider.shared.currentActiveUser.value else { throw APIService.APIError.explicit(.authenticationMissing) }
-        
-        let updatedProfile = try await APIService.shared.updateTabDisplaySettings(showFeaturedTab: editingViewModel.featuredTabVisibilitySetting == .showFeaturedTab, showMediaTab: editingViewModel.mediaTabVisibilitySetting == .showMediaTab, showMediaReplies: editingViewModel.mediaTabRepliesSetting == .includeMyRepliesToOthers, authenticationBox: authBox)
-        guard let updatedAccount = account?.byUpdatingTabSettings(
-            showFeaturedTab: updatedProfile.showFeatured,
-            showMediaTab: updatedProfile.showMedia,
-            showMediaReplies: updatedProfile.showMediaReplies
-        ) else { return }
-        
-        set(account: updatedAccount, relationship: .isMe, navigator: navigator)
-        editingViewModel.setAccount(updatedAccount)
-    }
-}
-
 extension ProfileViewModel {
     var bioIsEmpty: Bool {
         guard let bioForEdit = account?.bioForEdit else { return true }
@@ -1038,5 +481,6 @@ extension ProfileViewModel {
     }
 }
 
+let brandBackgroundColor: Color = Asset.Colors.Brand.lightBlurple.swiftUIColor.opacity(0.2)
 let avatarImageMaxSizeInPixels = CGSize(width: 400, height: 400)
 let bannerImageMaxSizeInPixels = CGSize(width: 1500, height: 500)
