+++ Mastodon/In Progress New Layout and Datamodel/Profile/ProfileView.swift
@@ -1000,173 +1000,7 @@ enum EditingStatus: Equatable {
     }
 }
 
-@MainActor
-@Observable class ProfileViewModel {
-    let relationshipViewModel = RelationshipViewModel()
-    let uuid = UUID() // helpful for debugging
-    
-    let editingViewModel = {
-        ProfileEditingViewModel()
-    }()
-    
-    var featuredHashtagsModel = FeaturedHashtagsModel()
-    
-    var editingStatus: EditingStatus = .cannotEdit
-    
-    struct HandleDetails {
-        let username: String
-        let domain: String
-        let isMyDomain: Bool
-    }
-    var navigator: MastodonNavigationRouter?
-    var account: MastodonAccount?
-    var relationship: MastodonAccount.Relationship?
-    var familiarFollowersViewModel: TimelineListViewModel?
-    var pagesToShow: [ProfilePage] = []
-    var postsViewModel: TimelineListViewModel? {
-        didSet {
-            Task {
-                switch await postsViewModel?.timeline {
-                case .userPosts(_, let queryFilter):
-                    activityFilter = queryFilter
-                default:
-                    activityFilter = nil
-                }
-            }
-        }
-    }
-    private var activityFilter: TimelineQueryFilter?
-    
-    var mediaViewModel: TimelineListViewModel?
-    {
-        didSet {
-            Task {
-                switch await mediaViewModel?.timeline {
-                case .userPosts(_, let queryFilter):
-                    mediaFilter = queryFilter
-                default:
-                    mediaFilter = nil
-                }
-            }
-        }
-    }
-    private var mediaFilter: TimelineQueryFilter?
-    var mediaViewAsyncRefresh: AsyncRefreshViewModel?
-    
-    var featuredItemsViewModel: TimelineListViewModel?
-    var featuredItemsAsyncRefresh: AsyncRefreshViewModel?
-    var selectedPage: ProfilePage = .activity
-    var focusedCustomField: Mastodon.Entity.Field?
-    var handleDetails: HandleDetails?
-    
-    var navigationButtons: [UIBarButtonItem] {
-        guard let account, let relationship else { return [] }
-        switch relationship {
-        case .isMe:
-            let settings = UIBarButtonItem(image: .init(systemName: "gearshape"), style: .plain, target: self, action: nil)
-            let hashtags = UIBarButtonItem(image: .init(systemName: "number"), style: .plain, target: self, action: nil)
-            let bookmarks = UIBarButtonItem(image: .init(systemName: "bookmark"), style: .plain, target: self, action: nil)
-            let favourites = UIBarButtonItem(image: .init(systemName: "star"), style: .plain, target: self, action: nil)
-            let share = UIBarButtonItem(image: .init(systemName: "square.and.arrow.up"), style: .plain, target: self, action: nil)
-            return [hashtags, bookmarks, favourites, share, settings]
-        case .isNotMe:
-            let menu = UIBarButtonItem(image: .init(systemName: "ellipsis.circle"), style: .plain, target: self, action: nil)
-            let reply = UIBarButtonItem(image: .init(systemName: "arrow.turn.up.left"), style: .plain, target: self, action: nil)
-            return [menu, reply]
-        }
-    }
-    
-    private var updateSubscription: AnyCancellable?
-    init() {
-        self.updateSubscription = FeedCoordinator.shared.$mostRecentUpdate
-            .receive(on: DispatchQueue.main)
-            .sink { [weak self] update in
-                guard let self, let update else { return }
-                self.incorporateUpdate(update)
-            }
-    }
-    
-    public func set(account: MastodonAccount, relationship: MastodonAccount.Relationship, navigator: MastodonNavigationRouter) {
-        self.navigator = navigator
-        self.account = account
-        self.editingViewModel.setAccount(account)
-        self.relationship = relationship
-        self.postsViewModel = TimelineListViewModel(timeline: .userPosts(userID: account.id, queryFilter: .init(.userPosts(featuredHashtagsModel))), navigator: navigator, asyncRefreshViewModel: AsyncRefreshViewModel())
-        
-        Task {
-            do {
-                try await featuredHashtagsModel.fetchFeaturedTags(account: account)
-                editingViewModel.updateFeaturedHashtags(featuredHashtagsModel.featuredHashtags)
-            } catch {
-                navigator.didReceiveError(error)
-            }
-        }
-        
-        if account.metadata.showsMediaTab {
-            if mediaViewModel == nil {
-                mediaViewAsyncRefresh = AsyncRefreshViewModel()
-                let mediaQueryFilter = TimelineQueryFilter(.mediaOnly)
-                mediaQueryFilter.excludeReplies = !account.metadata.mediaTabIncludesReplies
-                self.mediaViewModel = TimelineListViewModel(timeline: .userPosts(userID: account.id, queryFilter: mediaQueryFilter), navigator: navigator, asyncRefreshViewModel: mediaViewAsyncRefresh!)
-            } else {
-                self.updateMediaFilter()
-            }
-        } else {
-            self.mediaViewModel = nil
-        }
-        
-        if account.metadata.showsFeaturedTab {
-            if self.featuredItemsViewModel == nil {
-                featuredItemsAsyncRefresh = AsyncRefreshViewModel()
-                self.featuredItemsViewModel = TimelineListViewModel(timeline: .featuredItems(userID: account.id), navigator: navigator, asyncRefreshViewModel: featuredItemsAsyncRefresh!)
-            }
-        } else {
-            self.featuredItemsViewModel = nil
-        }
-        
-        self.relationshipViewModel.prepareForDisplay(relationship: relationship, theirAccountIsLocked: account.locked)
-        
-        self.relationshipViewModel.actionHandler = self.postsViewModel
-        
-        switch relationship {
-        case .isMe:
-            self.editingStatus = .notEditing
-        case .isNotMe:
-            self.editingStatus = .cannotEdit
-        }
-        
-        pagesToShow = pagesToShow(forAccount: account)
-        
-        Task {
-            let handle = account.handle
-            let handleComponents = handle.split(separator: "@").map { String($0) }
-            if handleComponents.count > 1 {
-                // server is included
-                self.handleDetails = .init(username: handleComponents.first ?? "", domain: handleComponents.last ?? "", isMyDomain: false)
-            } else {
-                // this account is on my server
-                if let myDomain = AuthenticationServiceProvider.shared.currentActiveUser.value?.domain {
-                    self.handleDetails = .init(username: handleComponents.first ?? "", domain: myDomain, isMyDomain: true)
-                }
-            }
-        }
-    }
-    
-    public func resetEditingViewModel() {
-        guard let account else { return }
-        editingViewModel.setAccount(account)
-    }
-    
-    public func updateMediaFilter() {
-        guard let mediaFilter, let account else { return }
-        let accountIsExcludingRepliesInMediaTab = !account.metadata.mediaTabIncludesReplies
-        guard mediaFilter.excludeReplies != accountIsExcludingRepliesInMediaTab else { return }
-        mediaFilter.excludeReplies = accountIsExcludingRepliesInMediaTab
-        Task {
-            await mediaViewModel?.forceReload(.mediaFilterUpdated)
-        }
-    }
-}
+
 
 extension ProfileViewModel: FeedCoordinatorUpdatable {
     func incorporateUpdate(_ update: UpdatedElement) {
@@ -1350,20 +1184,6 @@ extension ProfileViewModel {
         MastodonMenuAction.menuButton(systemImageName: action.iconSystemName, text: action.labelText) {
             self.handleAction(action)
         }
-    }
-}
-
-extension ProfileViewModel {
-    private func pagesToShow(forAccount account: MastodonAccount?) -> [ProfilePage] {
-        var pages = [ProfilePage.activity]
-        if account?.metadata.showsMediaTab == true {
-            pages.append(.mediaOnly)
-        }
-        let canShowFeaturedTab = AuthenticationServiceProvider.shared.currentActiveUser.value?.authentication.instanceConfiguration?.isAvailable(.featuredAccounts) ?? false
-        if canShowFeaturedTab && account?.metadata.showsFeaturedTab == true {
-            pages.append(.featured)
-        }
-        return pages
     }
 }
